refactor: bookmark feature - move directory & refactoring codes
