+++ src/components/Bookmark/BookmarkGrid/index.ts
@@ -1,0 +1,3 @@
+import BookmarkGrid from './BookmarkGrid';
+
+export default BookmarkGrid;
