+++ src/components/Bookmark/BookmarkHeader/index.ts
@@ -1,0 +1,3 @@
+import BookmarkHeader from './BookmarkHeader';
+
+export default BookmarkHeader;
