+++ src/components/Bookmark/BookmarkItem/index.ts
@@ -1,0 +1,3 @@
+import BookmarkItem from './BookmarkItem';
+
+export default BookmarkItem;
