--- src/pages/BookmarkPage/BookmarkPage.style.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledBookmark = styled.section`
-  display: flex;
-  justify-content: center;
-  padding-top: 16px;
-`;
