+++ src/pages/BookmarkPage/index.ts
@@ -1,0 +1,3 @@
+import BookmarkPage from './BookmarkPage';
+
+export default BookmarkPage;
