+++ src/pages/BookmarkPage/styles.ts
@@ -1,5 +1,11 @@
 import styled from 'styled-components';
 
+export const StyledBookmark = styled.section`
+  display: flex;
+  justify-content: center;
+  padding-top: 16px;
+`;
+
 export const StyledBookmarkContainer = styled.section`
   width: 100%;
   height: 100%;
