+++ src/utils/hooks/useBookmark/index.ts
@@ -1,0 +1,3 @@
+import useBookmark from './useBookmark';
+
+export default useBookmark;
